Convert our custom helper behavior into a component method
This removes the GeoblacklightHelperBehavior that was created
to generate our search result preview descriptions and makes it
into a tested component method instead, so that there's less
overriding of Blacklight.
