+++ app/components/geoblacklight/search_result_component.html.erb
@@ -35,11 +35,10 @@
       </span>
     <% end %>
   <% end %>
-    <div class='more-info-area border-bottom'>
+  <div class='more-info-area border-bottom'>
     <div id="doc-<%= @document.id %>-fields-collapse" class='collapse'>
       <small itemprop="description">
-        <%=  helpers.geoblacklight_present(:index_fields_display, @document) %>
+        <%= index_fields_display %>
       </small>
     </div>
-    </div>
-
+  </div>
