+++ app/components/geoblacklight/search_result_component.rb
@@ -9,5 +9,20 @@ module Geoblacklight
   #    Otherwise view_component will not provide the count value when calling the component.
   #
   class SearchResultComponent < Blacklight::DocumentComponent
+    # Presents configured index fields in search results. Passes values through
+    # configured helper_method. Multivalued fields separated by presenter
+    # field_value_separator (default: comma). Fields separated by period.
+    # @return [String]
+    def index_fields_display
+      fields_values = []
+      @presenter.configuration.index_fields.each do |_field_name, field_config|
+        val = @presenter.field_value(field_config).to_s
+        if val.present?
+          val += "." unless val.end_with?(".")
+          fields_values << val
+        end
+      end
+      safe_join(fields_values, " ")
+    end
   end
 end
