+++ app/helpers/blacklight_helper.rb
@@ -2,7 +2,6 @@
 
 module BlacklightHelper
   include Blacklight::BlacklightHelperBehavior
-  include Geoblacklight::GeoblacklightHelperBehavior
   include Blacklight::CatalogHelperBehavior
 
   def render_document_sidebar_partial(document = @document)
