+++ lib/generators/geoblacklight/templates/catalog_controller.rb
@@ -52,10 +52,6 @@ class CatalogController < ApplicationController
     config.show.sidebar_component = Geoblacklight::SidebarComponent
     config.header_component = Geoblacklight::HeaderComponent
 
-    ##
-    # Configure the index document presenter.
-    config.index.document_presenter_class = Geoblacklight::DocumentPresenter
-
     # solr fields that will be treated as facets by the blacklight application
     #   The ordering of the field names is the order of the display
     #
